Refactor WebtoonTranslationsView
