+++ app/src/main/java/eu/kanade/translation/presentation/Utils.kt
@@ -1,0 +1,14 @@
+package eu.kanade.translation.presentation
+
+import android.util.DisplayMetrics
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+
+
+@Composable
+fun Float.pxToDp(): Dp {
+    val density = LocalDensity.current.density
+    return (this / (density / DisplayMetrics.DENSITY_DEFAULT)).dp
+}
